requirements.txt
asgiref==3.2.10
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
folium==0.11.0
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
dj-database-url==0.5.0
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
whitenoise==5.2.0
pycrypto==2.6.1
